+++ enoslib/infra/enos_vmong5k/configuration.py
@@ -14,6 +14,8 @@ class Configuration(BaseConfiguration):
 
     def __init__(self):
         super().__init__()
+        self.gateway = None
+        self.gateway_user = None
         self.job_name = DEFAULT_JOB_NAME
         self.queue = DEFAULT_QUEUE
         self.walltime = DEFAULT_WALLTIME